Fix build warnings
src/dbdump.c: In function ‘foldprint’:
src/dbdump.c:35:5: error: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 2 has type ‘long long unsigned int’ [-Werror=format]
src/dbdump.c:35:5: error: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 3 has type ‘long long unsigned int’ [-Werror=format]
src/dbdump.c:35:5: error: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 4 has type ‘long long unsigned int’ [-Werror=format]

src/dbinfo.c: In function ‘process_file’:
src/dbinfo.c:57:14: error: variable ‘datasize’ set but not used [-Werror=unused-but-set-variable]

Reviewed-on: http://review.couchbase.org/14298
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>

+++ src/dbdump.c
@@ -32,7 +32,7 @@ static int foldprint(Db *db, DocInfo *docinfo, void *ctx)
     printf("     id: ");
     printsb(&docinfo->id);
     printf("     content_meta: %d\n", docinfo->content_meta);
-    printf("     cas: %"PRIu64", expiry: %"PRIu64", flags: %"PRIu64"\n", (long long unsigned int) cas, (long long unsigned int) expiry, (long long unsigned int) flags);
+    printf("     cas: %"PRIu64", expiry: %"PRIu32", flags: %"PRIu32"\n", cas, expiry, flags);
     if (docinfo->deleted) {
         printf("     doc deleted\n");
     }
